controller implementations

+++ src/main/java/com/tahadonuk/restaurantmanagementsystem/service/CustomerService.java
@@ -1,7 +1,8 @@
 package com.tahadonuk.restaurantmanagementsystem.service;
 
-import com.tahadonuk.restaurantmanagementsystem.data.entity.user.customer.Customer;
+import com.tahadonuk.restaurantmanagementsystem.data.entity.user.Customer;
 import com.tahadonuk.restaurantmanagementsystem.data.repository.CustomerRepository;
+import com.tahadonuk.restaurantmanagementsystem.exception.NotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -16,22 +17,25 @@ public class CustomerService {
         customerRepo.save(customer);
     }
 
-    public boolean isExists(long id) {
-        return customerRepo.existsById(id);
-    }
-
-    public Customer getById(long id) {
+    public Customer getById(long id) throws NotFoundException {
         if(isExists(id)) {
             return customerRepo.findById(id).get();
         }
-        else return null;
+        else throw new NotFoundException("No such order with given ID");
     }
 
-    public void deleteCustomer(long id) {
-        customerRepo.deleteById(id);
+    public void deleteCustomer(long id) throws NotFoundException {
+        if(isExists(id)) {
+            customerRepo.deleteById(id);
+        } else throw new NotFoundException("No such order with given ID");
     }
 
     public List<Customer> getAll() {
         return customerRepo.findAll();
+    }
+
+
+    public boolean isExists(long id) {
+        return customerRepo.existsById(id);
     }
 }
